[SCRUM-29] refactor: command / query 서비스로 분리
