+++ src/main/java/com/projectlyrics/server/domain/artist/service/ArtistQueryService.java
@@ -1,0 +1,8 @@
+package com.projectlyrics.server.domain.artist.service;
+
+import com.projectlyrics.server.domain.artist.dto.response.GetArtistResponse;
+
+public interface ArtistQueryService {
+
+  GetArtistResponse getArtist(Long artistId);
+}
