+++ src/main/java/com/projectlyrics/server/domain/artist/service/impl/ArtistCommandServiceImpl.java
@@ -3,10 +3,9 @@ package com.projectlyrics.server.domain.artist.service.impl;
 import com.projectlyrics.server.domain.artist.dto.request.AddArtistRequest;
 import com.projectlyrics.server.domain.artist.dto.request.UpdateArtistRequest;
 import com.projectlyrics.server.domain.artist.dto.response.AddArtistResponse;
-import com.projectlyrics.server.domain.artist.dto.response.GetArtistResponse;
 import com.projectlyrics.server.domain.artist.dto.response.UpdateArtistResponse;
 import com.projectlyrics.server.domain.artist.repository.CommandQueryArtistRepository;
-import com.projectlyrics.server.domain.artist.service.ArtistService;
+import com.projectlyrics.server.domain.artist.service.ArtistCommandService;
 import com.projectlyrics.server.global.error_code.ErrorCode;
 import com.projectlyrics.server.global.exception.BusinessException;
 import java.time.Clock;
@@ -15,21 +14,19 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 @Service
-public class ArtistServiceImpl implements ArtistService {
+public class ArtistCommandServiceImpl implements ArtistCommandService {
 
   private final CommandQueryArtistRepository commandArtistRepository;
   private final Clock clock;
 
-  @Transactional
   @Override
   public AddArtistResponse addArtist(AddArtistRequest request) {
     var savedArtist = commandArtistRepository.save(request.toEntity());
     return AddArtistResponse.of(savedArtist.getId());
   }
 
-  @Transactional
   @Override
   public UpdateArtistResponse updateArtist(Long artistId, UpdateArtistRequest request) {
     var artist = commandArtistRepository.findByIdAndNotDeleted(artistId)
@@ -42,7 +39,6 @@ public class ArtistServiceImpl implements ArtistService {
     return UpdateArtistResponse.from(artist);
   }
 
-  @Transactional
   @Override
   public void deleteArtist(Long artistId) {
     var artist = commandArtistRepository.findByIdAndNotDeleted(artistId)
@@ -50,13 +46,5 @@ public class ArtistServiceImpl implements ArtistService {
 
     // TODO: 인증 구현되면 deletedById 값 수정
     artist.getCommonField().delete(1L, clock.systemDefaultZone());
-  }
-
-  @Override
-  public GetArtistResponse getArtist(Long artistId) {
-    var artist = commandArtistRepository.findByIdAndNotDeleted(artistId)
-        .orElseThrow(() -> new BusinessException(ErrorCode.ARTIST_NOT_FOUND));
-
-    return GetArtistResponse.from(artist);
   }
 }
