+++ src/test/java/com/projectlyrics/server/domain/artist/service/ArtistCommandServiceTest.java
@@ -1,37 +1,34 @@
 package com.projectlyrics.server.domain.artist.service;
 
-import com.projectlyrics.server.domain.artist.dto.request.AddArtistRequest;
-import com.projectlyrics.server.domain.artist.dto.request.UpdateArtistRequest;
-import com.projectlyrics.server.domain.artist.entity.Artist;
-import com.projectlyrics.server.domain.artist.repository.CommandQueryArtistRepository;
-import com.projectlyrics.server.domain.artist.service.impl.ArtistServiceImpl;
-import com.projectlyrics.server.domain.common.entity.enumerate.EntityStatusEnum;
-import com.projectlyrics.server.global.exception.BusinessException;
-import com.projectlyrics.server.utils.ArtistTestUtil;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.util.Optional;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.catchThrowable;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
-import static org.mockito.Mockito.doReturn;
+
+import com.projectlyrics.server.domain.artist.dto.request.AddArtistRequest;
+import com.projectlyrics.server.domain.artist.dto.request.UpdateArtistRequest;
+import com.projectlyrics.server.domain.artist.entity.Artist;
+import com.projectlyrics.server.domain.artist.repository.CommandQueryArtistRepository;
+import com.projectlyrics.server.domain.artist.service.impl.ArtistCommandServiceImpl;
+import com.projectlyrics.server.domain.common.entity.enumerate.EntityStatusEnum;
+import com.projectlyrics.server.global.exception.BusinessException;
+import com.projectlyrics.server.utils.ArtistTestUtil;
+import java.util.Optional;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-class ArtistServiceTest {
+class ArtistCommandServiceTest {
 
   @InjectMocks
-  private ArtistServiceImpl sut;
+  private ArtistCommandServiceImpl sut;
 
   @Mock
   private CommandQueryArtistRepository commandArtistRepository;
@@ -106,25 +103,6 @@ class ArtistServiceTest {
     assertThat(artist.getCommonField().getDeletedAt()).isNotNull();
     assertThat(artist.getCommonField().getDeletedBy()).isNotNull();
     assertThat(artist.getCommonField().getStatus()).isEqualTo(EntityStatusEnum.DELETED);
-  }
-
-  @Test
-  void 아티스트의_PK를_전달받아_아티스트_데이터를_조회해_반환한다() {
-    // given
-    Long artistId = 1L;
-    var artist = Mockito.spy(ArtistTestUtil.create());
-    given(commandArtistRepository.findByIdAndNotDeleted(artistId)).willReturn(Optional.of(artist));
-    doReturn(artistId).when(artist).getId();
-
-    // when
-    var getArtistResponse = sut.getArtist(artistId);
-
-    // then
-    then(commandArtistRepository).should().findByIdAndNotDeleted(anyLong());
-    assertThat(getArtistResponse.id()).isEqualTo(artistId);
-    assertThat(getArtistResponse.name()).isEqualTo(artist.getName());
-    assertThat(getArtistResponse.englishName()).isEqualTo(artist.getEnglishName());
-    assertThat(getArtistResponse.profileImageCdnLink()).isEqualTo(artist.getProfileImageCdnLink());
   }
 
   private AddArtistRequest createAddArtistRequest() {
